[ REFACTOR ] handleCodeShareButtonOnClick 함수 내 로직을 따로 분리하여 copyToClipboard 함수로 구현 (#51)
Why
- 프로젝트와 상관없이 사용될 수 있는 로직은 유틸함수로 분리해야 한다.

What
- dom.js 내 param으로 들어온 값을 클립보드에 복사해주는 유틸함수를 구현하였다.
